Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0          # SQLite async driver (dev only)
asyncpg==0.29.0            # PostgreSQL async driver (prod)
alembic==1.13.2
pydantic==2.9.2
# celery[redis]==5.4.0
# redis==5.0.8
python-dotenv==1.0.1
httpx==0.27.2              # for calling WhatsApp Cloud API
